backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
python-jose
passlib[bcrypt]
python-multipart
Pillow
torch
transformers
pgvector
face_recognition
scikit-learn